[ts-sdk] add modular exports for keypairs
