+++ .changeset/quick-monkeys-hang.md
@@ -1,0 +1,5 @@
+---
+'@mysten/sui.js': minor
+---
+
+Add keypair exports to allow modular imports for various keypair types
