+++ sdk/typescript/keypairs/ed25519/package.json
@@ -1,0 +1,5 @@
+{
+	"private": true,
+	"import": "../dist/esm/keypairs/ed25519/index.js",
+	"main": "../dist/cjs/keypairs/ed25519/index.js"
+}
