+++ sdk/typescript/src/cryptography/publickey.ts
@@ -1,10 +1,6 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import { Ed25519PublicKey } from './ed25519-publickey.js';
-import { Secp256k1PublicKey } from './secp256k1-publickey.js';
-import { SignatureScheme } from './signature.js';
-
 /**
  * Value to be converted into public key.
  */
@@ -58,14 +54,4 @@ export interface PublicKey {
 	 * Return signature scheme flag of the public key
 	 */
 	flag(): number;
-}
-
-export function publicKeyFromSerialized(schema: SignatureScheme, pubKey: string): PublicKey {
-	if (schema === 'ED25519') {
-		return new Ed25519PublicKey(pubKey);
-	}
-	if (schema === 'Secp256k1') {
-		return new Secp256k1PublicKey(pubKey);
-	}
-	throw new Error('Unknown public key schema');
 }
