+++ sdk/typescript/src/cryptography/signature.ts
@@ -1,16 +1,8 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import { fromB64, toB64 } from '@mysten/bcs';
-import { Ed25519PublicKey } from './ed25519-publickey.js';
-import { PublicKey } from './publickey.js';
-import { Secp256k1PublicKey } from './secp256k1-publickey.js';
-import { Secp256r1PublicKey } from './secp256r1-publickey.js';
-import { decodeMultiSig } from './multisig.js';
+import type { PublicKey } from './publickey';
 
-/**
- * A keypair used for signing transactions.
- */
 export type SignatureScheme = 'ED25519' | 'Secp256k1' | 'Secp256r1' | 'MultiSig';
 
 /**
@@ -45,61 +37,3 @@ export const SIGNATURE_FLAG_TO_SCHEME = {
 } as const;
 
 export type SignatureFlag = keyof typeof SIGNATURE_FLAG_TO_SCHEME;
-
-export function toSerializedSignature({
-	signature,
-	signatureScheme,
-	pubKey,
-}: SignaturePubkeyPair): SerializedSignature {
-	const serializedSignature = new Uint8Array(1 + signature.length + pubKey.toBytes().length);
-	serializedSignature.set([SIGNATURE_SCHEME_TO_FLAG[signatureScheme]]);
-	serializedSignature.set(signature, 1);
-	serializedSignature.set(pubKey.toBytes(), 1 + signature.length);
-	return toB64(serializedSignature);
-}
-
-/// Expects to parse a serialized signature by its signature scheme to a list of signature
-/// and public key pairs. The list is of length 1 if it is not multisig.
-export function toParsedSignaturePubkeyPair(
-	serializedSignature: SerializedSignature,
-): SignaturePubkeyPair[] {
-	const bytes = fromB64(serializedSignature);
-	const signatureScheme =
-		SIGNATURE_FLAG_TO_SCHEME[bytes[0] as keyof typeof SIGNATURE_FLAG_TO_SCHEME];
-
-	if (signatureScheme === 'MultiSig') {
-		return decodeMultiSig(serializedSignature);
-	}
-
-	const SIGNATURE_SCHEME_TO_PUBLIC_KEY = {
-		ED25519: Ed25519PublicKey,
-		Secp256k1: Secp256k1PublicKey,
-		Secp256r1: Secp256r1PublicKey,
-	};
-
-	const PublicKey = SIGNATURE_SCHEME_TO_PUBLIC_KEY[signatureScheme];
-
-	const signature = bytes.slice(1, bytes.length - PublicKey.SIZE);
-	const pubkeyBytes = bytes.slice(1 + signature.length);
-	const pubKey = new PublicKey(pubkeyBytes);
-
-	return [
-		{
-			signatureScheme,
-			signature,
-			pubKey,
-		},
-	];
-}
-
-/// Expects to parse a single signature pubkey pair from the serialized
-/// signature. Use this only if multisig is not expected.
-export function toSingleSignaturePubkeyPair(
-	serializedSignature: SerializedSignature,
-): SignaturePubkeyPair {
-	const res = toParsedSignaturePubkeyPair(serializedSignature);
-	if (res.length !== 1) {
-		throw Error('Expected a single signature');
-	}
-	return res[0];
-}
