+++ sdk/typescript/src/keypairs/ed25519/publickey.ts
@@ -3,9 +3,9 @@
 
 import { blake2b } from '@noble/hashes/blake2b';
 import { fromB64, toB64 } from '@mysten/bcs';
-import { bytesEqual, PublicKeyInitData } from './publickey.js';
-import { SIGNATURE_SCHEME_TO_FLAG } from './signature.js';
-import { normalizeSuiAddress, SUI_ADDRESS_LENGTH } from '../types/index.js';
+import { bytesEqual, PublicKeyInitData } from '../../cryptography/publickey.js';
+import { SIGNATURE_SCHEME_TO_FLAG } from '../../cryptography/signature.js';
+import { normalizeSuiAddress, SUI_ADDRESS_LENGTH } from '../../types/index.js';
 import { bytesToHex } from '@noble/hashes/utils';
 
 const PUBLIC_KEY_SIZE = 32;
