+++ sdk/typescript/src/keypairs/secp256k1/index.ts
@@ -1,0 +1,5 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+export * from './keypair.js';
+export * from './publickey.js';
