+++ sdk/typescript/src/keypairs/secp256k1/publickey.ts
@@ -4,9 +4,9 @@
 import { fromB64, toB64 } from '@mysten/bcs';
 import { blake2b } from '@noble/hashes/blake2b';
 import { bytesToHex } from '@noble/hashes/utils';
-import { normalizeSuiAddress, SUI_ADDRESS_LENGTH } from '../types/index.js';
-import { bytesEqual, PublicKey, PublicKeyInitData } from './publickey.js';
-import { SIGNATURE_SCHEME_TO_FLAG } from './signature.js';
+import { normalizeSuiAddress, SUI_ADDRESS_LENGTH } from '../../types/index.js';
+import { bytesEqual, PublicKey, PublicKeyInitData } from '../../cryptography/publickey.js';
+import { SIGNATURE_SCHEME_TO_FLAG } from '../../cryptography/signature.js';
 
 const SECP256K1_PUBLIC_KEY_SIZE = 33;
