+++ sdk/typescript/src/keypairs/secp256r1/keypair.ts
@@ -1,15 +1,15 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import type { ExportedKeypair, Keypair } from './keypair.js';
-import { PublicKey } from './publickey.js';
+import type { ExportedKeypair, Keypair } from '../../cryptography/keypair.js';
+import { PublicKey } from '../../cryptography/publickey.js';
 import { sha256 } from '@noble/hashes/sha256';
-import { Secp256r1PublicKey } from './secp256r1-publickey.js';
+import { Secp256r1PublicKey } from './publickey.js';
 import { secp256r1 } from '@noble/curves/p256';
-import { isValidBIP32Path, mnemonicToSeed } from './mnemonics.js';
+import { isValidBIP32Path, mnemonicToSeed } from '../../cryptography/mnemonics.js';
 import { HDKey } from '@scure/bip32';
 import { toB64 } from '@mysten/bcs';
-import { SignatureScheme } from './signature.js';
+import type { SignatureScheme } from '../../cryptography/signature.js';
 import { bytesToHex } from '@noble/hashes/utils';
 import { blake2b } from '@noble/hashes/blake2b';
