+++ sdk/typescript/src/utils/verify.ts
@@ -6,8 +6,9 @@ import nacl from 'tweetnacl';
 import { IntentScope, messageWithIntent } from './intent.js';
 import { secp256k1 } from '@noble/curves/secp256k1';
 import { sha256 } from '@noble/hashes/sha256';
-import { SerializedSignature, toSingleSignaturePubkeyPair } from '../cryptography/signature.js';
+import { SerializedSignature } from '../cryptography/signature.js';
 import { blake2b } from '@noble/hashes/blake2b';
+import { toSingleSignaturePubkeyPair } from '../cryptography/utils.js';
 
 // TODO: This might actually make sense to eventually move to the `Keypair` instances themselves, as
 // it could allow the Sui.js to be tree-shaken a little better, possibly allowing keypairs that are
