+++ sdk/typescript/test/unit/cryptography/secp256k1-publickey.test.ts
@@ -3,7 +3,7 @@
 
 import { toB64, toHEX } from '@mysten/bcs';
 import { describe, it, expect } from 'vitest';
-import { Secp256k1PublicKey } from '../../../src/cryptography/secp256k1-publickey';
+import { Secp256k1PublicKey } from '../../../src/keypairs/secp256k1/publickey';
 import { INVALID_SECP256K1_PUBLIC_KEY, VALID_SECP256K1_PUBLIC_KEY } from './secp256k1-keypair.test';
 
 // Test case generated against CLI:
