+++ sdk/typescript/test/unit/cryptography/secp256r1-publickey.test.ts
@@ -3,7 +3,7 @@
 
 import { toB64, toHEX } from '@mysten/bcs';
 import { describe, it, expect } from 'vitest';
-import { Secp256r1PublicKey } from '../../../src/cryptography/secp256r1-publickey';
+import { Secp256r1PublicKey } from '../../../src/keypairs/secp256r1/publickey';
 import { INVALID_SECP256R1_PUBLIC_KEY, VALID_SECP256R1_PUBLIC_KEY } from './secp256r1-keypair.test';
 
 // Test case generated against CLI:
